Add day 12 solutions
The initial solution was summing the count of pots with plants, not the
sum of the indices, so I was always off by a bunch.

Then I messed up a bunch because I was summing on generation 19 instead
of 20 which I fixed by running the sample input and checking that I got
the right answer (I didn't) and analyzing why. I was a generation shy.

After that, the second part was obviously impossible to calculate
directly, so I found the pattern that kicked in around generation 125
and jumped to generation 200 for round numbers and found the indices of
the plants at that point and translated the indices upward on the line
of plants to the right generation. There's your answer!