Added day 14 solutions
This one seemed much easier to plug away at than day 13. Less work to
build the initial model. Searching for a number sequence in the existing
vector was the trickiest thing, but given that the recipe list grows
fairly slowly, you only need to check the last few entries for a match
(I checked the last ten just to be safe.)

On a side note, the first implementation of the initial `test_recipes`
function created new vectors for recipes and elves and it was way way
slower, so just another reminder that rampant allocation can hurt you.